+++ src/main/java/com/netcracker/skillstable/model/dto/Department.java
@@ -4,15 +4,15 @@ import java.util.HashSet;
 import java.util.Set;
 
 public class Department extends OrgItem {
-    private static final Long entTypeId = 4L;
+    private static final Integer entTypeId = 4;
     private Set<Team> teams = new HashSet<>();
-    private static final Long teamRefId = 4L;
+    private static final Integer teamRefId = 4;
 
 
     public Department() {
     }
 
-    public Department(Long id, String name) {
+    public Department(Integer id, String name) {
         super(id, name);
     }
 
@@ -26,11 +26,11 @@ public class Department extends OrgItem {
     }
 
 
-    public static Long getEntTypeId() {
+    public static Integer getEntTypeId() {
         return entTypeId;
     }
 
-    public static Long getTeamRefId() {
+    public static Integer getTeamRefId() {
         return teamRefId;
     }
